+++ lib/__tests__/index.js
@@ -3,7 +3,7 @@
 var assert = require('assert');
 var del = require('del');
 var fs = require('fs');
-var metalToolsSoy = require('../index');
+var metalToolsSoy = require('../../index');
 var sinon = require('sinon');
 var vfs = require('vinyl-fs');
