+++ lib/pipelines/__tests__/compileSoy.js
@@ -2,7 +2,7 @@
 
 var assert = require('assert');
 var childProcess = require('child_process');
-var compileSoy = require('../../../lib/pipelines/compileSoy');
+var compileSoy = require('../compileSoy');
 var del = require('del');
 var ignore = require('gulp-ignore');
 var path = require('path');
